sdk: sendMessageToContact
