--- js-miniapp-sdk/src/chat-service.ts
@@ -1,17 +1,0 @@
-import {
-  MiniAppBridge,
-  MessageToContact,
-} from '../../js-miniapp-bridge/src';
-
-/** @internal */
-function getBridge() {
-  // tslint:disable:no-any
-  return (window as any).MiniAppBridge as MiniAppBridge;
-}
-
-export class ChatService {
-  
-  sendMessageToContact(message: MessageToContact): Promise<string | undefined> {
-    return getBridge().sendMessageToContact(message);
-  }
-}
