+++ js-miniapp-sdk/src/miniapp.ts
@@ -1,5 +1,4 @@
 import {
-  MiniAppBridge,
   Reward,
   DevicePermission,
   CustomPermission,
@@ -8,15 +7,10 @@ import {
   CustomPermissionStatus,
   ShareInfoType,
   ScreenOrientation,
-  AccessTokenData,
-  Contact,
 } from '../../js-miniapp-bridge/src';
-
-/** @internal */
-function getBridge() {
-  // tslint:disable:no-any
-  return (window as any).MiniAppBridge as MiniAppBridge;
-}
+import { UserInfoProvider, UserInfo } from './modules/user-info';
+import { ChatService } from './modules/chat-service';
+import { getBridge } from './utils';
 
 /**
  * A module layer for webapps and mobile native interaction.
@@ -116,59 +110,9 @@ interface Platform {
   getPlatform(): string;
 }
 
-/**
- * Interfaces to retrieve User profile related information.
- */
-export interface UserInfoProvider {
-  /**
-   * Fetches the username from host app.
-   * You should request the {@link CustomPermissionName.USER_NAME} permission before using this method.
-   * @returns Username saved in the host app user profile.
-   */
-  getUserName(): Promise<string>;
-
-  /**
-   * Fetches the profile photo URI from host app.
-   * You should request the {@link CustomPermissionName.PROFILE_PHOTO} permission before using this method.
-   * @returns Profile photo saved in the host app user profile.
-   */
-  getProfilePhoto(): Promise<string>;
-
-  /**
-   * Fetches the contact list from host app.
-   * You should request the {@link CustomPermissionName.CONTACT_LIST} permission before using this method.
-   * @returns Contact list in the host app user profile.
-   */
-  getContacts(): Promise<Contact[]>;
-
-  /**
-   * Fetches the access token from host app.
-   * @returns Access token from native host app.
-   */
-  getAccessToken(): Promise<AccessTokenData>;
-}
-
-/** @internal */
-class UserInfo implements UserInfoProvider {
-  getUserName(): Promise<string> {
-    return getBridge().getUserName();
-  }
-
-  getProfilePhoto(): Promise<string> {
-    return getBridge().getProfilePhoto();
-  }
-
-  getContacts(): Promise<Contact[]> {
-    return getBridge().getContacts();
-  }
-
-  getAccessToken(): Promise<AccessTokenData> {
-    return getBridge().getAccessToken();
-  }
-}
-
 export class MiniApp implements MiniAppFeatures, Ad, Platform {
   user: UserInfoProvider = new UserInfo();
+  chatService = new ChatService();
 
   private requestPermission(permissionType: DevicePermission): Promise<string> {
     return getBridge().requestPermission(permissionType);
