+++ js-miniapp-sdk/src/utils.ts
@@ -1,0 +1,7 @@
+import { MiniAppBridge } from '../../js-miniapp-bridge/src';
+
+/** @internal */
+export function getBridge() {
+  // tslint:disable:no-any
+  return (window as any).MiniAppBridge as MiniAppBridge;
+}
